+++ packages/pocket-ic/CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Functions `PocketIc::from_config`, `PocketIc::from_config_and_max_request_time`, and `PocketIc::from_config_and_server_url`.
   Use the `PocketIcBuilder` instead.
 
+### Changed
+- The type `Topology` becomes a struct with two fields: `subnet_configs` contains an association of subnet ids to their configurations
+  and `default_effective_canister_id` contains a default effective canister id for canister creation.
+
 
 
 ## 5.0.0 - 2024-09-12